Client will now send chosen character to host.
Host will ensure that only one of each character is chosen.
Client cannot send multiple ready signals.
If a character is unselected that player is no longer ready, and the character can be chosen by another player.
Modified UICharacterSelect so ui could be reset without an onClickListener.
Fixed issue in connection lobby with playerID moving to int from long.

+++ Source/Cypher/app/src/main/java/com/semaphore_soft/apps/cypher/ConnectionLobbyActivity.java
@@ -180,7 +180,7 @@ public class ConnectionLobbyActivity extends AppCompatActivity implements Respon
                                Toast.LENGTH_SHORT).show();
                 Intent intent = new Intent(getBaseContext(), CharacterSelectActivity.class);
                 intent.putExtra("host", host);
-                intent.putExtra("player", (long) 0);
+                intent.putExtra("player", 0);
                 intent.putExtra("numClients", playerID);
                 startActivity(intent);
                 break;
